Refactor `StickerSource` into `sticker` module
